save user while creating job

+++ backend/job/views.py
@@ -1,4 +1,4 @@
-from rest_framework.decorators import api_view
+from rest_framework.decorators import api_view, permission_classes
 from .models import Job
 from .serializers import JobSerializer
 from rest_framework.response import Response
@@ -7,6 +7,7 @@ from rest_framework import status
 from django.db.models import Avg, Min, Max, Count
 from .filters import JobsFilter
 from rest_framework.pagination import PageNumberPagination
+from rest_framework.permissions import IsAuthenticated
 
 # Create your views here.
 
@@ -41,7 +42,9 @@ def getJob(request, pk):
 
 
 @api_view(['POST'])
+@permission_classes([IsAuthenticated])
 def newJob(request):
+    request.data['user'] = request.user
     data = request.data
     job = Job.objects.create(**data)
 
@@ -50,8 +53,12 @@ def newJob(request):
 
 
 @api_view(['PUT'])
+@permission_classes([IsAuthenticated])
 def newJob(request, pk):
     job = get_object_or_404(Job, id=pk)
+
+    if job.user != request.user:
+        return Response({'message': 'You can not update this job'}, status=status.HTTP_403_FORBIDDEN)
 
     job.title = request.data['title']
     job.description = request.data['description']
@@ -72,8 +79,11 @@ def newJob(request, pk):
 
 
 @api_view(['DELETE'])
+@permission_classes([IsAuthenticated])
 def deleteJob(request, pk):
     job = get_object_or_404(Job, id=pk)
+    if job.user != request.user:
+        return Response({'message': 'You can not delete this job'}, status=status.HTTP_403_FORBIDDEN)
     job.delete()
     return Response({'message': f'Job with {job.email} is Deleted.'}, status=status.HTTP_200_OK)
